--- TempoWorld/Source/TempoObservableEvents/Private/TempoObservableEvents.cpp
@@ -1,17 +1,0 @@
-﻿#include "TempoObservableEvents.h"
-
-#define LOCTEXT_NAMESPACE "FTempoObservableEventsModule"
-
-void FTempoObservableEventsModule::StartupModule()
-{
-    
-}
-
-void FTempoObservableEventsModule::ShutdownModule()
-{
-    
-}
-
-#undef LOCTEXT_NAMESPACE
-    
-IMPLEMENT_MODULE(FTempoObservableEventsModule, TempoObservableEvents)
