--- TempoWorld/Source/TempoObservableEvents/Public/TempoObservableEvents.h
@@ -1,11 +1,0 @@
-﻿#pragma once
-
-#include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
-
-class FTempoObservableEventsModule : public IModuleInterface
-{
-public:
-    virtual void StartupModule() override;
-    virtual void ShutdownModule() override;
-};
