+++ TempoWorld/Source/TempoWorld/TempoWorld.Build.cs
@@ -2,7 +2,7 @@
 
 using UnrealBuildTool;
 
-public class TempoWorld : ModuleRules
+public class TempoWorld : TempoModuleRules
 {
     public TempoWorld(ReadOnlyTargetRules Target) : base(Target)
     {
@@ -11,7 +11,11 @@ public class TempoWorld : ModuleRules
         PublicDependencyModuleNames.AddRange(
             new string[]
             {
+                // Unreal
                 "Core",
+                // Tempo
+                "TempoCoreShared",
+                "TempoScripting",
             }
         );
 
@@ -20,6 +24,7 @@ public class TempoWorld : ModuleRules
             {
                 "CoreUObject",
                 "Engine",
+                "TempoCore",
             }
         );
     }
