+++ TempoWorld/TempoWorld.uplugin
@@ -23,11 +23,6 @@
 			"Name": "TempoMapQuery",
 			"Type": "Runtime",
 			"LoadingPhase": "Default"
-		},
-		{
-			"Name": "TempoObservableEvents",
-			"Type": "Runtime",
-			"LoadingPhase": "Default"
 		}
 	],
 	"Plugins": [
